Consolidate 'home' route into 'events' route
